Se agrego la opcion de ver y agregar beneficiarios

+++ src/models/persona_models/persona_beneficiario_models.rs
@@ -5,5 +5,5 @@ pub struct BeneficiarioPersonaModelo {
     pub id_persona_beneficiario: i32,
     pub id_persona: i64,
     pub entre_calle: i64,
-    pub y_calle: i32,
+    pub y_calle: i64,
 }
